requirements.txt
pandas==2.2.0
psycopg2-binary==2.9.9
streamlit==1.31.0
faker==22.6.0
plotly==5.18.0
redis==5.0.1
